shorts-to-analysis/clustering/requirements.txt
# Clustering and ML Libraries
faiss-cpu>=1.7.4
openai>=1.3.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Additional dependencies that may be needed
numpy>=1.24.0
pandas>=2.0.0
